+++ CHANGELOG.md
@@ -6,11 +6,13 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [4.0.1] - 2022-06-29
 - Added TLS capability
+- Added ability to override postfix and use mailx command.
 
 ### Fixed
 - chk_mst_log: Fixed problem of appending chk_slv function data to SlaveLogs list.
 
 ### Changed
+- data_out: Added option to send_mail call to override postfix and use mailx command.
 - config/slave.txt.TEMPLATE: Added TLS entry.
 - Documentation updates.